fix(#716): move NatsEventBus from core/ to services/ — concrete driver does not belong in kernel
NatsEventBus is a concrete NATS JetStream driver implementation, not a kernel ABC.
Per KERNEL-ARCHITECTURE.md, drivers are config-time DI implementations that belong
outside the static kernel layer. Moved to nexus.services.event_bus_nats and updated
all imports across source and test files.
