+++ tests/e2e/nats/test_nats_event_bus_e2e.py
@@ -168,7 +168,7 @@ def server_app():
     nexus_fs._audit_strict_mode = False
 
     # Verify event bus is NATS
-    from nexus.core.event_bus_nats import NatsEventBus
+    from nexus.services.event_bus_nats import NatsEventBus
 
     assert nexus_fs._event_bus is not None
     assert isinstance(nexus_fs._event_bus, NatsEventBus)
@@ -224,7 +224,7 @@ class TestServerStartup:
 
     def test_event_bus_is_nats(self, client, nexus_fs):
         """Event bus should be NatsEventBus."""
-        from nexus.core.event_bus_nats import NatsEventBus
+        from nexus.services.event_bus_nats import NatsEventBus
 
         assert isinstance(nexus_fs._event_bus, NatsEventBus)
